Fix security + performance issues from code review
Security:
- save_set: validate exercise belongs to the log (IDOR fix)
- substitute_exercise: catch ValueError/TypeError on bad exercise_id
- weekly_plan: change start_session link to POST form (was 405)

Performance:
- Streak calculation: single query for last 60 days instead of N loops
- Volume trend: batch fetch all logs once instead of per-plan queries
- total_volume_kg: use .all() loop instead of .filter() that breaks prefetch

Model:
- target_rpe: PositiveIntegerField → DecimalField(3,1) to support RPE 7.5

Other:
- Dashboard progress bar uses dynamic width from consistency_pct
- Remove unused `from copy import copy` in service.py

+++ plans/models.py
@@ -60,7 +60,7 @@ class PlannedExercise(models.Model):
     sets = models.PositiveIntegerField()
     reps_min = models.PositiveIntegerField()
     reps_max = models.PositiveIntegerField()
-    target_rpe = models.PositiveIntegerField()
+    target_rpe = models.DecimalField(max_digits=3, decimal_places=1)
     rest = models.CharField(max_length=20)
     weight_kg = models.FloatField(null=True, blank=True)  # null = establish_needed
